Argus timer library
This adds a library of routines for maintaining a large number of
timers.  This is designed to be usable from the ArgusClientTimeout()
function.  A "hashed wheel" with a configurable number of slots
(teeth?) holds references to the timers.  Each slot has a sorted
tree of pointers to timer structures, ordered by expiration time.

When a timer is inserted into the wheel, a slot is chosen by finding
the expiration time modulo the number of slots.  The timer structure
is added to the balanced tree in that slot.

Every time the wheel is advanced (once per ArgusClientTimeout() for
example) the timer tree in the current slot is searched for timers
that expire sometime during the current period.  Each timer specifies
a callback function which is invoked at this time and the timer
structure is removed from the tree.

Once every revolution of the timer wheel, the wheel's time is synced
with the operating systems's time.  If the wheel got little behind,
fire all of the timers that would have been missed because of jumping
to the new time.  This happens some on systems that don't support
clock_nanosleep() (with absolute times).

A timer is stopped by simply removing it from its tree.

For systems that support clock_nanosleep(), use the monotonic clock
and absolute times to keep argus_timer's internal "clock"
(argus_timer_wheel->now) from drifting too far.