modify argus_output to use "ArgusWireFmtBuffer"s
Queue buffers as they will be sent on the network instead of queueing
flow records.  This avoids the need to serialize each buffer once
for each client for each transmit attempt.

Instead of putting the queue-header structure inside of ArgusWireFmtBuffer,
create a separate structure to hold the queue/list entry and a
pointer to the ArgusWireFmtBuffer.  Doing so allows the same buffer
to be added to multiple queues.  In this case, the buffer can be
queued separately for each socket waiting to transmit.

Trac ticket #56