radhcp: adjust non-lease timer to reset if not bound
If the non-lease timer is running and we receive another packet for
that transaction, stop the timer if transitioning to the BOUND state
and reset the timer otherwise.  Prevents a partial transaction from
hanging around indefinitely in some cases.

Avoid race between RabootpProtoTimersNonleaseSet() which can reset
the timer and __discover_exp_cb() which is called when the timer
expires.