radhcp: temporary fix to lock ordering
The dhcp parser aquires a lock for the current transaction and then
runs callbacks based on state change, etc.  The callbacks can start
and stop timers, which aquires the timer wheel lock.

Timer callbacks are run from the timer thread with the timer wheel
lock aquired, but some of the callbacks aquire a transaction lock.

To avoid possible deadlock, ensure that these two locks are always
aquired in the same order.  Since the timer callbacks will ALWAYS
be run with the timer wheel lock aquired, make the parser code match
-- aquire the timer lock and then run __parse_one_dhcp_record()
which aquires a transaction lock.

This cannot be a long-term solution because it results in the timer
lock being held for a relatively long period of time which will
interfere with time-keeping.