per-listen-fd version num in ArgusOutputStruct
Associate a protocol version with each listen file descriptor instead
of with each output processor.  Each output thread can handle multiple
formats.  This will cause clients to choose a version implicitly by
connecting to a particular tcp port, rather than (a) explicitly for V5
clients and (b) by default for V3 clients.

Store a per-listen-fd version number array in ArgusOutputStruct and
ArgusParserStruct.  Also store proto version in each ArgusClientData
structure.  This allows the version to be checked when a new connection
is accepted and any time data is sent to a client.