radhcp: add rabootp_lease_pullup.[ch]
Some implementations of the DHCP client do not renew leases as allowed
by the protocol, but instead aquire new leases.  This means that the
transaction ID changes even though the assigned address does not in
most cases.  Viewing a history of leases for a host that behaves this
way is difficult and may be misleading.  For the purposes of displaying
lease times, provide a function that will take overlapping or consecutive
leases that result in the same address assignment and "pull them up"
into a single lease.