update interval-tree subtreehi on node update
When a lease is extended, the corresponding node in the interval tree
must be updated to reflect the lease's new end time.  This can change
the max value for interval end times in that subtree.  When that
changes the new value must be checked against that of parent nodes
since they may also need to be updated.

Fixes: ca1009777377 ("radhcp: update interval tree")