radhcp: rabootp_querystr.[ch] - parse query strings
Add a slightly more sophisticated (and generic) parser for query strings.
This parses options in the form "<lhs>["="<rhs>]".  A description of the
options is passed into the function RabootpParseQueryString and an array
with pointers to the consituent strings is passed back.  No conversion
to numeric types is done by the parser - only string manipulation.