radhcp: time for sql table name can be specified
Introduce ArgusCreateSQLSaveTableName_usec() which accepts a time value
given in microseconds which will be passed to strftime() to format
the date section of the table name IF the flow record parameter (ns)
is NULL.  If ns is NULL and the usec value is zero or less, fall back
to the original NULL-ns behavior.

ArgusCreateSQLSaveTableName is now a simple wrapper around
ArgusCreateSQLSaveTableName_usec.