find system perl extenstion dir from configure
There seems to be little consistency in locations for perl extensions
among operating systems.  Solaris, OSX and RHEL/Fedora all have
different default @INC paths for perl, for example.

Use the autoconf test for perl that is found in autoconf-archive.
The --with-perl configure option is removed with this change but
the location of the perl binary can still be specified on the
configure commandline by setting the PERL variable.

  % ./configure PERL=/bin/perl

For linux RPMs, use the perl vendor library directory determined by
rpmbuild.  This will change the location in which QoSient perl modules
are installed by argus-clients* RPMs to /usr/share/perl5/vendor_perl.