radhcp: treat slightly non-overlapping leases as consecutive
There are cases where the client may renew (or in the case of windows,
aquire a new but identical lease) at the very last moment. Just because
of delays due to port mirroring, queueing in software or even a tardy
DHCP participant, radhcp may see the two leases as non-overlapping by as
little as 1/10 - 1/20 of a second. Add a slop factor to allow for this
particular brand of stupidity. Leases that miss overlapping by less than
a second should now be considered overlapping and will be pulled-up into
a single lease when generating a summary.

Trac ticket #109