argus_timer, radhcp: rescheduled timer uses same struct argus_timer
If rescheduling a timer, do not free the original timer struct; reuse
it instead.  There is no way in the timer API to pass back an updated
argus_timer pointer when the timer is rescheduled during handling of
its expiry.

In radhcp clear the lease timer pointer when its holddown expires,
not when the original lease expires.  The pointer is still good now.

Also avoid a race condition where a lease that was relinquished early
has started a new holddown timer.