radhcp: add sanity checks to removal functions
When removing a lease from the interval or client trees --
ArgusDhcpIntvlTreeRemove() and ArgusDhcpClientTreeRemove(), a pointer
to the target ArgusDhcpStruct is passed to the removal function.
If a node in the tree is found, compare that node's ArgusDhcpStruct
pointer to the one passed in by the caller.  Only remove the node
if those pointers match.