radns: do not process SOA requests
Print SOA records when seen in the stream, but don't add CNAME or PTR
response data, when the DNS request was for an SOA record.  Add double
quotes to address and name string printing, not just for json output.
Modify relts_print parameter to be unsigned int from int (conform to
tcpdump).