specify -S localhost only once to rasqlinsert
Most of the files used by the rasqlinsert@.service template include
a source (-S) option to be passed to rasqlinsert.  This shouldn't
be done since source selection is handled by the service file.  The
result is that each rasqlinsert attaches to radium twice.

Remove the -S option from ${datadir}/argus-clients/rasqlinsert.*.