radhcp: allow search for address prefixes
A query string can now contain a netmask in the addr= field.  For
example
  SEARCH: when=2017/02/15+1d,addr=192.168.8.0/24,pullup

Reverse the notion of left and right in the patricia  tree traversal
functions to match the order of data stored in the tree.  Stop tree
traveral when the next node is the parent of the subtree root.

Trac ticket #124