separate file list from ArgusInput into dedicated structure
Introduce a new structure, ArgusFileInput, to hold lists of files.
This avoids allocating an ArgusInput structure for each filename in
the list.  ArgusInput is rather large (658520 bytes on x86_64 linux)
and for a list of thousands of files can result in gigabytes of memory
being allocated.  This is especially a problem on 32-bit platforms since
the address space available for the heap may only be a couple of GB.

Use the new structure in argus_main.c.  Allocate one ArgusInput structure
and reuse it for each file.

Additional commits will follow that update the various clients that
examine the list of files directly.

Trac ticket #128