win7: enable ramanage archive compress & delete
Add the configuration items to ramanage.conf necessary to compress
anything after 0 days (that isn't already compressed) and delete any
files in the archive that are older than 180 days.  Also create the
/staging directory from the installer.

The policy is applied even to the file after the -r option, so if a
value greater than 0 is provided for RAMANAGE_RPOLICY_COMPRESS_AFTER,
the -r file is not compressed.  This may need to change.