add client commandline option to specify v3 output files
Introduce the -3 option for clients that are able to write files.
This will cause the output files to contain argus v3 (open source)
formatted records.  Most clients can now be used to translate from
one argus version to the other.  Previously, only radium could do
this and required the user to use a tcp connection to read the v3
data.

Trac ticket #134