rahisto: print json list if multiple metrics
If more than one -H option is supplied and JSON mode is enabled, wrap
the JSON objects in an array.

Also output the name of the metric for each histogram, regardless of
JSON mode.

When only a single metric is computed the output is identical to previous
versions of rahisto.