rahisto-updatesql: use new perflow mode in rahisto
rahisto_update_index_table() and rahisto_update_values_table() in
rahisto.pm will look for the "instance" object in the hash decoded from
rahisto's json output.  rahisto is called with the flow model specified
as "saddr" and, as such, the instance object/dictionary will have only
an saddr field.  If present, use this field to set the address and mask
columns in the two SQL tables that are updated on histogram insertion.
Otherwise use the values passed in via the parameter list.

This avoids running rahisto, with a filter, once for each address in the
source file.  It is instead run once and outputs a (set of) histogram(s)
for each source IP address.