common: fix label string formatting in RaLocalityLabel
RaLocalityLabel() appends the locality label to the global char
RaLocalityLabelBuffer[256] every time its called.  In other words,
it start formatting the label at the beginning of the buffer once
per process.  This caused the buffer to be overrun, strlen() to
report values that later caused arithmetic underflow, etc., and
eventually corrupted the stack.

Start formatting the label string at the beginning of the buffer
each time RaLocalityLabel() is called.  Do not rely on strlen() for
string length since we get this from snprintf().

Trac ticket #145