ramanage: esacpe or convert the filename passed to curl
When compiled without libcurl support, ramanage will attempt to run an
external curl binary to upload a file.  On cygwin builds this filename
needs to be converted to a Windows/DOS path.  On non-cygwin builds the
filename may need to be escaped with backslashes to protect it from the
bourne shell.