ramanage: clear curl commandline string for each file
Be sure to reset the string length and remaining byte count each time
a new curl commandline is formatted.  Previously, __upload_init() did
this once before an entire batch of files was uploaded.  Modify that
function to allow it to be called multiple times in the same process
and call once for each file uploaded.