common: fix warnings about truncated output in ArgusPrintCountryCode
GCC 7 reports the following:
  ./argus_util.c:22245:37: warning: ‘__builtin_snprintf’ output may be truncated before the last format character [-Wformat-truncation=]
                     snprintf (buf, 3, "%s", caddr->cco);
                                     ^
  ./argus_util.c:22245:19: note: ‘__builtin_snprintf’ output between 1 and 4 bytes into a destination of size 3
                     snprintf (buf, 3, "%s", caddr->cco);
                   ^

caddr->cco is an array of four chars, the last of which must be NULL.
The size argument of snprintf includes room for the terminating NULL,
so increment the size by one.

  ./argus_util.c:22257:25: warning: ‘__builtin_snprintf’ output truncated before the last format character [-Wformat-truncation=]
         snprintf (buf, 2, "  ");
                         ^
  ./argus_util.c:22257:7: note: ‘__builtin_snprintf’ output 3 bytes into a destination of size 2
         snprintf (buf, 2, "  ");

Fix this truncation warning by replacing snprintf with sprintf --
there are no format specifiers to expand and we don't know the length
of buf anyways.