common: fix warning for misleading indentation in ArgusGrepUserData
GCC 7 reports:
  ./argus_grep.c: In function ‘ArgusGrepUserData’:
  ./argus_grep.c:155:13: warning: this ‘if’ clause does not guard... [-Wmisleading-indentation]
               if ((retn = ArgusGrepBuf (&parser->upreg[i], buf, &buf[len])))
               ^~
  ./argus_grep.c:157:16: note: ...this statement, but the latter is misleadingly indented as if it were guarded by the ‘if’
                  break;
                  ^~~~~

Assume that the current behavior is wrong and that the loop should be
terminated when the pattern is found, instead of always being terminated
after the first iteration.  Also replace auto-increment with an assignment
to one; there is no need to first read the contents of that variable.