ramanage: tell curl binary return non-zero for http errors
The --fail option will cause the curl(1) program to return 22 if the HTTP
response indicates an error.  Add this option to the curl commandline
when libcurl is not present so that protocol failures (authentication,
premissions, etc.) will cause the process to terminate abnormally and
ramanage(1) will not move the source file into the staging directory.