ArgusClientTimeout not called often enough for file inputs
Only gettimeofday is used to determine when to call ArgusClientTimeout
and, in rabins at least, gettimeofday is used inside of ArgusClientTimeout
to determine if the bins should be processed.

When input is NOT a live stream, the "time of day" can only be determined
by the flow record start times.  Use these times to decide when bins
should be processed/flushed for file inputs.

Consolodate calls to ArgusClientTimeout() in ArgusHandleRecord, since
this is the one common point for all input data types - live, argus file,
netflow, jflow, etc.