ramanage: skip missing files
It is possible for multiple instances of ramanage(1) to have some of
the same files in their lists of inputs.  One instance of ramanage(1)
may process a file before another instance gets around to it.  This is
not a problem and is to be expected at times where concurrent execution
takes place.  Instead of complaining, skip the missing file and move on
to the next.