ramanage: fix file mtime comparison in RamanageTrimFiles
The __compare_argus_input_file_mtime() function is intended to be used
with qsort and an array of pointers.  As a result, the two arguments,
as passed in by qsort(), are pointers to pointers.  RamanageTrimFiles()
also uses this comparison function, but passes the wrong data type.
Fix RamanageTrimFiles() to pass pointers to pointers.