make sorting optional in RaProcessRecursiveFiles
The current/expected behavior of argus clients is to process all files
from the command line, including those found by a recursive directory
search, in alphanumeric order of filenames.

ramanage(1) needs to process the list of files in order of modification
time.  Add a parameter to RaProcessRecursiveFiles() that will cause
it to skip sorting the files since for ramanage this could be a lot of
unnecessary extra work.