common: keep the results of stat when building file list
Commit 5512db91e1c4 changed RaProcessRecursiveFiles to copy the buffer
filled in by stat() into the ArgusInputFile structure for the newly
added file.  Applications needing information about the file do not have
to call stat() a second time.

Make the same change in RaDescend().  stat() is called for every file
added to the input list in this function; always keep the results so
that all entries in the list have this info.

ArgusParseArgs() can also add files to the list for the -r option.
Add a call to stat() here.

Fixes: 5512db91e1c4 ("RaProcessRecursiveFiles: keep the results of stat()")