ramanage: do not stat files in input list
As of commit 61b7c98a8734 ("common: keep the results of stat when building
file list"), it is no longer necessary to check for files that have not
been passed to stat().  Only stat the primary input file ("exemplar"
in the source) after the random delay and global lock aquisition in case
it was removed in the mean time.