common: fix connecting to live source after reading stdin
Connecting to a live stream after reading stdin will fail because the
client is unable to send the START: message.

ArgusWriteConnection() checks to make sure that it is not attempting to
write data to std{in,ou,err} by testing the ArgusInput structure's fd
field against the range [0-2], unless the daemon option (-d) was provided.

If stdio is read first and then closed, the zero file descriptor becomes
available and is returned by the next call to socket().  This confuses
the test in ArgusWriteConnection causing it to write nothing and the
connection appears to have been closed.

The problem can be reproduced by running this command:
  % bin/ra -X -r - -S ::1

and pressing ctrl-d (or piping in the output of another command, etc.).

The fileno() function will return -1 if the file has been closed.
Compare fd to fileno(stdin), etc., instead of a hard-coded range.  Also,
this may help portability since it is not guaranteed that an operating
system will use 0, 1 and 2 for stdio.