make all dhcp SQL table names start with dhcp_
Update the scripts run by the nightly systemd timers to use tables in the
dhcpFlows database that start with the prefix "dhcp_".  This means that
the scripts create and use tables with names like dhcp_summary_2018_12_30
instead of summary_2018_12_30.  This naming convention ensures that the
table names are unique from the perspective of any one sensor UUID.
This is necessary for when the SQL schema later changes to per-UUID
databases,

Include a migration script, migration_from_5.0-3.4.rhel.23_dhcp, which
will rename all of the existing tables in the dhcpFlows database for
consistency with per-UUID databases.